simplify VertexAttribInfo
enforce value count restrictions and simplify size checking

+++ Projects/SFGenericModel/GenericMesh.cs
@@ -230,7 +230,7 @@ namespace SFGenericModel
                 // -1 means not found, which is usually a result of the attribute being unused.
                 int index = shader.GetVertexAttributeUniformLocation(attribute.name);
                 if (index != -1)
-                    GL.VertexAttribPointer(index, attribute.valueCount, attribute.vertexAttribPointerType, false, vertexSizeInBytes, offset);
+                    GL.VertexAttribPointer(index, (int)attribute.valueCount, attribute.vertexAttribPointerType, false, vertexSizeInBytes, offset);
                 // Move offset to next attribute.
                 offset += attribute.sizeInBytes;
             }
